wsd: test: move test file-serving logic into own home
